+++ tests/Js/Main/ComparisonTests.fs
@@ -373,6 +373,7 @@ let tests =
 
     testCase "max works with primitives" <| fun () ->
         max 1 2 |> equal 2
+        max 1m 2m |> equal 2m
         Math.Max(1, 2) |> equal 2
         max "a" "b" |> equal "b"
 
@@ -388,6 +389,7 @@ let tests =
 
     testCase "min works with primitives" <| fun () ->
         min 1 2 |> equal 1
+        min 1m 2m |> equal 1m
         Math.Min(1, 2) |> equal 1
         min "a" "b" |> equal "a"